requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
pyTelegramBotAPI==4.14.0
gunicorn==21.2.0
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23 